microservice-combined-approach/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
gitpython==3.1.36
python-magic==0.4.27
tiktoken==0.5.1
tenacity==8.2.3
faiss-cpu==1.7.4
numpy==1.25.2
pydantic==2.3.0
torch>=2.0.0
transformers==4.38.2
openai>=1.0.0,<2.0.0